# docs/swagger.yaml
basePath: /
definitions:
...
  portainer_api.RequestInput:
    properties:
      username:
        example: exampleusername
        type: string
    type: object
  utils.BadRequest:
    properties:
      error:
        example: string
        type: string
      message:
        example: message
        type: string
      result:
        example: "null"
        type: string
      status:
        example: 400
        type: integer
      success:
        example: false
        type: boolean
    type: object
  utils.FoundFail:
    properties:
      error:
        example: message
        type: string
      message:
        example: message
        type: string
      result:
        example: "null"
        type: string
      status:
        example: 302
        type: integer
      success:
        example: false
        type: boolean
    type: object
  utils.FoundSuccess:
    properties:
      error:
        example: "null"
        type: string
      message:
        example: message
        type: string
      result:
        example: message
        type: string
      status:
        example: 302
        type: integer
      success:
        example: true
        type: boolean
    type: object
  utils.InternalServerError:
    properties:
      error:
        example: message
        type: string
      message:
        example: message
        type: string
      result:
        example: "null"
        type: string
      status:
        example: 500
        type: integer
      success:
        example: false
        type: boolean
    type: object
  utils.Success:
    properties:
      error:
        example: "null"
        type: string
      message:
        example: message
        type: string
      result:
        example: message
        type: string
      status:
        example: 200
        type: integer
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: Documentation for Automate Deployment Restful API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Automate Deployment API
  version: "1.0"
paths:
...
  /api/v1/system/is-not-exists:
    post:
      consumes:
      - application/json
      description: True if stack is not exist
      parameters:
      - description: Body Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/portainer_api.RequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/utils.Success'
        "302":
          description: Found
          schema:
            $ref: '#/definitions/utils.FoundFail'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.BadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.InternalServerError'
      security:
      - X-Token: []
      summary: True if stack is not exist
      tags:
      - System
...
  /api/v1/system/status:
    post:
      consumes:
      - application/json
      description: Get status of stack
      parameters:
      - description: Body Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/portainer_api.RequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/utils.Success'
        "302":
          description: Found
          schema:
            $ref: '#/definitions/utils.FoundSuccess'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.BadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.InternalServerError'
      security:
      - X-Token: []
      summary: Get status of stack
      tags:
      - System
securityDefinitions:
  X-Token:
    description: Input your token authorized
    in: header
    name: X-Token
    type: apiKey
swagger: "2.0"